# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Go Funny API
  version: 0.2.0
  summary: Backend API contract for Go Funny trip collaboration and settlement.
  description: |
    This spec covers the app's business APIs.
    Better Auth endpoints under `/api/auth/*` are intentionally excluded because the frontend uses the Better Auth client directly.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://go-funny-backend.onrender.com
    description: Production server
tags:
  - name: Trips
  - name: Members
  - name: Expenses
  - name: Contributions
  - name: Settlement
  - name: Invitations
security:
  - cookieAuth: []

paths:
  /trips:
    get:
      tags: [Trips]
      operationId: getTrips
      summary: Get trips visible to the current user
      responses:
        '200':
          description: Trip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Trips]
      operationId: createTrip
      summary: Create a trip and join it as the first member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /trips/{tripId}:
    get:
      tags: [Trips]
      operationId: getTripById
      summary: Get trip detail if the current user is a member
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Trip detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Trips]
      operationId: updateTrip
      summary: Update trip metadata
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTripRequest'
      responses:
        '200':
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Trips]
      operationId: deleteTrip
      summary: Delete a trip
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '204':
          description: Trip deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /members/{tripId}:
    get:
      tags: [Members]
      operationId: getTripMembers
      summary: Get trip members
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /members/{tripId}/{memberId}:
    delete:
      tags: [Members]
      operationId: deleteTripMember
      summary: Delete a trip member
      parameters:
        - $ref: '#/components/parameters/TripId'
        - $ref: '#/components/parameters/MemberId'
      responses:
        '204':
          description: Member deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /expenses/{tripId}:
    get:
      tags: [Expenses]
      operationId: getTripExpenses
      summary: Get trip expenses
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Expense list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Expenses]
      operationId: createTripExpense
      summary: Create an expense
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /expenses/{tripId}/{expenseId}:
    delete:
      tags: [Expenses]
      operationId: deleteTripExpense
      summary: Delete an expense
      parameters:
        - $ref: '#/components/parameters/TripId'
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '204':
          description: Expense deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /contributions/{tripId}:
    get:
      tags: [Contributions]
      operationId: getTripContributions
      summary: Get trip contributions
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Contribution list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contribution'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Contributions]
      operationId: createTripContribution
      summary: Create a contribution
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContributionRequest'
      responses:
        '201':
          description: Contribution created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contribution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /settlement/{tripId}:
    get:
      tags: [Settlement]
      operationId: getTripSettlement
      summary: Get settlement result
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Settlement result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /invitations/{tripId}:
    post:
      tags: [Invitations]
      operationId: createTripInvitation
      summary: Create an invitation link for a trip
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /invitations/{token}:
    get:
      tags: [Invitations]
      operationId: getInvitationByToken
      summary: Get invitation info by token
      security: []
      parameters:
        - $ref: '#/components/parameters/InvitationToken'
      responses:
        '200':
          description: Invitation info with trip data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationWithTrip'
        '404':
          $ref: '#/components/responses/NotFound'

  /invitations/{token}/accept:
    post:
      tags: [Invitations]
      operationId: acceptInvitation
      summary: Accept an invitation and join the trip
      parameters:
        - $ref: '#/components/parameters/InvitationToken'
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripMembership'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Better Auth session cookie

  parameters:
    TripId:
      name: tripId
      in: path
      required: true
      schema:
        type: string
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
    ExpenseId:
      name: expenseId
      in: path
      required: true
      schema:
        type: string
    InvitationToken:
      name: token
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    TripMode:
      type: string
      enum: [expense, pool]

    SplitType:
      type: string
      enum: [equal_all, equal_selected, custom]

    InvitationRole:
      type: string
      enum: [editor]

    User:
      type: object
      required: [id, email, name, emailVerified, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        emailVerified:
          type: boolean
        image:
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    TripMembership:
      type: object
      required: [id, tripId, userId, createdAt]
      properties:
        id:
          type: string
        tripId:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time

    Member:
      allOf:
        - $ref: '#/components/schemas/TripMembership'
        - type: object
          required: [user]
          properties:
            user:
              $ref: '#/components/schemas/User'

    ExpenseSplit:
      type: object
      required: [id, expenseId, membershipId, amount, createdAt]
      properties:
        id:
          type: string
        expenseId:
          type: string
        membershipId:
          type: string
        amount:
          type: integer
        createdAt:
          type: string
          format: date-time

    Expense:
      type: object
      required: [id, tripId, title, amount, date, splitType, createdAt]
      properties:
        id:
          type: string
        tripId:
          type: string
        title:
          type: string
        amount:
          type: integer
        date:
          type: string
          format: date-time
        splitType:
          $ref: '#/components/schemas/SplitType'
        payerMembershipId:
          type:
            - string
            - 'null'
        note:
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time
        splits:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseSplit'

    Contribution:
      type: object
      required: [id, tripId, membershipId, amount, date, createdAt]
      properties:
        id:
          type: string
        tripId:
          type: string
        membershipId:
          type: string
        amount:
          type: integer
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    SettlementTransfer:
      type: object
      required: [from, to, amount]
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: number

    Settlement:
      type: object
      required: [tripId, mode, transfers]
      properties:
        tripId:
          type: string
        mode:
          $ref: '#/components/schemas/TripMode'
        transfers:
          type: array
          items:
            $ref: '#/components/schemas/SettlementTransfer'

    Invitation:
      type: object
      required: [id, tripId, token, role, usedCount, createdAt]
      properties:
        id:
          type: string
        tripId:
          type: string
        token:
          type: string
        role:
          $ref: '#/components/schemas/InvitationRole'
        maxUses:
          type:
            - integer
            - 'null'
        usedCount:
          type: integer
        expiresAt:
          type:
            - string
            - 'null'
          format: date-time
        acceptedAt:
          type:
            - string
            - 'null'
          format: date-time
        revokedAt:
          type:
            - string
            - 'null'
          format: date-time
        createdByUserId:
          type:
            - string
            - 'null'
        createdAt:
          type: string
          format: date-time

    InvitationWithTrip:
      allOf:
        - $ref: '#/components/schemas/Invitation'
        - type: object
          required: [trip]
          properties:
            trip:
              $ref: '#/components/schemas/TripSummary'

    TripSummary:
      type: object
      required: [id, title, mode, startDate, endDate, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        location:
          type:
            - string
            - 'null'
        mode:
          $ref: '#/components/schemas/TripMode'
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    Trip:
      allOf:
        - $ref: '#/components/schemas/TripSummary'
        - type: object
          required: [memberships, expenses, contributions]
          properties:
            memberships:
              type: array
              items:
                $ref: '#/components/schemas/Member'
            expenses:
              type: array
              items:
                $ref: '#/components/schemas/Expense'
            contributions:
              type: array
              items:
                $ref: '#/components/schemas/Contribution'

    CreateTripRequest:
      type: object
      required: [title, mode, startDate, endDate]
      properties:
        title:
          type: string
        location:
          type:
            - string
            - 'null'
        mode:
          $ref: '#/components/schemas/TripMode'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    UpdateTripRequest:
      type: object
      properties:
        title:
          type: string
        location:
          type:
            - string
            - 'null'
        mode:
          $ref: '#/components/schemas/TripMode'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    CreateExpenseRequest:
      type: object
      required: [title, amount, splitType, date]
      properties:
        title:
          type: string
        amount:
          type: integer
          minimum: 0
        splitType:
          $ref: '#/components/schemas/SplitType'
        payerMembershipId:
          type:
            - string
            - 'null'
        date:
          type: string
          format: date
        note:
          type:
            - string
            - 'null'

    CreateContributionRequest:
      type: object
      required: [membershipId, amount, date]
      properties:
        membershipId:
          type: string
        amount:
          type: integer
          minimum: 0
        date:
          type: string
          format: date

    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string